Tidy and move validation options logic to CheckoutFieldRegistrationUtils utility
